% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OLSFunction.R
\name{OLS2}
\alias{OLS2}
\title{OLS Coefficients}
\usage{
OLS2(y, X)
}
\arguments{
\item{y}{\code{numeric} or \code{matrix}.
The left hand
side variable of the model to be estimated.}

\item{X}{\code{matrix}. Contains the right
hand side variables of the
model to be estimated. Must include a
column of 1's if you
wish to include an intercept.}
}
\value{
A \code{matrix} of linear regression
coefficients.
}
\description{
Calculates OLS coefficients
  using matrix methods.
}
\details{
Longer description.
}

